@@ -3,21 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^12.1.0",
-    "@testing-library/user-event": "^13.2.1",
-    "axios": "^0.24.0",
-    "chance": "^1.1.8",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.3.4",
+    "chance": "^1.1.10",
     "csvtojson": "^2.0.10",
     "isdev": "^1.0.1",
-    "moment-timezone": "^0.5.33",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "moment-timezone": "^0.5.41",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "react-slideshow-image": "^3.6.0",
-    "web-vitals": "^2.1.0"
+    "react-router-dom": "^6.8.2",
+    "react-scripts": "5.0.1",
+    "react-slideshow-image": "^4.2.1",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -55,13 +55,13 @@
   },
   "homepage": "https://growmoreinternational.com/",
   "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.25.2",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "gh-pages": "^3.2.3",
+    "eslint": "^8.35.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "gh-pages": "^5.0.0",
     "react-snap": "^1.23.0"
   }
 }
